+++ src/app/demo/pc/menu/inline/demo.component.ts
