+++ src/app/demo/pc/menu/inline/demo.module.ts
