+++ src/app/pc-demo-list.component.ts
@@ -55,7 +55,6 @@ import {routerConfig as dateTimePickerConfig} from "./demo/pc/date-time-picker/d
 import {routerConfig as rangeDateTimePickerConfig} from "./demo/pc/range-date-time-picker/demo-set.module";
 import {routerConfig as progressConfig} from "./demo/pc/progress/demo-set.module";
 import {routerConfig as colorSelectConfig} from "./demo/pc/color-select/demo-set.module";
-import {routerConfig as navigationMenuConfig} from "./demo/pc/navigation-menu/demo-set.module";
 import {routerConfigPC} from "./router-config";
 import {PopupPositionType} from "../jigsaw/common/service/popup.service";
 
@@ -205,7 +204,6 @@ export class DemoListManager {
         this._addRouterConfig(routerConfig, 'range-date-time-picker', rangeDateTimePickerConfig);
         this._addRouterConfig(routerConfig, 'progress', progressConfig);
         this._addRouterConfig(routerConfig, 'color-select', colorSelectConfig);
-        this._addRouterConfig(routerConfig, 'navigation-menu', navigationMenuConfig);
     }
 
     private static _addRouterConfig(routerConfig: any[], path: string, childConfig: any[]) {
