+++ src/jigsaw/pc-components/theming/all-theme.scss
@@ -58,4 +58,4 @@
 @import "../date-and-time/date-time-select";
 @import "../progress/progress";
 @import "../color-select/color-select";
-@import "../navigation-menu/navigation-menu";
+@import "../menu/navigation-menu";
