+++ src/jigsaw/public_api.ts
@@ -82,4 +82,3 @@ export * from "./pc-components/upload/index";
 export * from "./pc-components/viewport/viewport";
 export * from "./pc-components/menu/index";
 export * from "./pc-components/color-select/index";
-export * from "./pc-components/navigation-menu/index";
